requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.0
PyJWT==2.8.0
loguru==0.7.2
pydantic==2.5.2
python-dotenv==1.0.0